+++ PixelWalle/Assets/scr/AST/ExpresionesLogicas/Mayor.cs
@@ -9,7 +9,7 @@ public class Mayor : ExpresionBinaria
     {
         Right.Calculate();
         Left.Calculate();
-        if ((double)Left.Value > (double)Right.Value) Value = (bool)true;
+        if ((int)Left.Value > (int)Right.Value) Value = (bool)true;
         else Value = (bool)false;
     }
